stats_service/requirements.txt
aiokafka==0.12.0
alembic==1.14.0
annotated-types==0.7.0
async-timeout==5.0.1
asyncpg==0.30.0
attrs==24.2.0
autoflake==2.3.1
certifi==2024.8.30
charset-normalizer==3.4.0
confluent-kafka==2.4.0
exceptiongroup==1.2.2
greenlet==3.1.1
idna==3.10
iniconfig==2.0.0
jsonlines==4.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kafka-python==2.0.2
Mako==1.3.8
MarkupSafe==3.0.2
mock==5.1.0
orjson==3.10.12
packaging==24.2
pluggy==1.5.0
psycopg2-binary==2.9.10
pydantic==2.10.3
pydantic-settings==2.6.1
pydantic_core==2.27.1
pyflakes==3.2.0
pytest==8.3.4
pytest-asyncio==0.25.0
pytest-mock==3.14.0
python-dotenv==1.0.1
quixstreams==3.4.0
referencing==0.35.1
requests==2.32.3
rocksdict==0.3.24
rpds-py==0.22.3
SQLAlchemy==2.0.36
tomli==2.2.1
typing_extensions==4.12.2
urllib3==2.2.3
